requirements.txt
# Web框架
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# MySQL客户端
mysqlclient==2.2.1

# WSGI服务器
gunicorn==21.2.0

# 任务调度
django-apscheduler==0.6.2
APScheduler==3.10.4

# 工具包
python-dotenv==1.0.0
requests==2.31.0

# 开发工具
django-debug-toolbar==4.2.0 